requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
python-dotenv==0.21.1
Flask-Migrate==3.1.0
Flask-Reuploaded==1.2.0
psycopg2-binary==2.9.3
pytest==7.0.1
python-dotenv==0.21.1
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Werkzeug==2.2.3
gevent==22.10.2
mysqlclient==2.1.1
Flask-Admin==1.6.1
sqlalchemy-serializer==1.4.12
Flask-Login==0.6.3
alembic==1.13.1
python-docx==0.8.11
pyotp==2.9.0
Flask-Mail==0.10.0
sendgrid ==6.11.0
fpdf2==2.7.6